PB-90: Test out looks with tab version
